MunicipalitiesDict: move to dict
